Ported the benckmarks to JMH
