--- src/main/java/com/mitchellbosecke/pebble/benchmark/Benchmark.java
@@ -1,9 +1,0 @@
-package com.mitchellbosecke.pebble.benchmark;
-
-public interface Benchmark {
-	
-	public void setup();
-
-	public void run(int reps);
-
-}
