--- src/main/java/com/mitchellbosecke/pebble/benchmark/CompilationsBenchmark.java
@@ -1,35 +1,0 @@
-package com.mitchellbosecke.pebble.benchmark;
-
-import com.mitchellbosecke.pebble.PebbleEngine;
-import com.mitchellbosecke.pebble.error.PebbleException;
-import com.mitchellbosecke.pebble.template.PebbleTemplate;
-
-public class CompilationsBenchmark implements Benchmark {
-
-    private PebbleEngine engine;
-
-    private final String templateName = "listing";
-
-    public void run(int reps) {
-
-        @SuppressWarnings("unused")
-        PebbleTemplate template;
-
-        for (int i = 0; i < reps; i++) {
-            try {
-                template = engine.getTemplate(templateName);
-            } catch (PebbleException e) {
-                throw new RuntimeException(e);
-            }
-        }
-    }
-
-    @Override
-    public void setup() {
-        this.engine = new PebbleEngine();
-        engine.getLoader().setPrefix("templates");
-        engine.getLoader().setSuffix(".html");
-        engine.setTemplateCache(null);
-    }
-
-}
